fix bug

+++ core/api/src/main/java/org/onosproject/net/flow/criteria/Criteria.java
@@ -656,155 +656,155 @@ public final class Criteria {
         return new Mac_SrcCriterion(mac_src, mask);
     }
 
-    public static Criterion selectVlan1_Tpid(short vlan1_tpid) {
+    public static Criterion selectVlan1_Tpid(long vlan1_tpid) {
         return new Vlan1_TpidCriterion(vlan1_tpid);
     }
   
-    public static Criterion selectVlan1_Tpid(short vlan1_tpid, short mask) {
+    public static Criterion selectVlan1_Tpid(long vlan1_tpid, long mask) {
         return new Vlan1_TpidCriterion(vlan1_tpid, mask);
     }
 
-    public static Criterion selectVlan1_Qid(short vlan1_qid) {
+    public static Criterion selectVlan1_Qid(long vlan1_qid) {
         return new Vlan1_QidCriterion(vlan1_qid);
     }
   
-    public static Criterion selectVlan1_Qid(short vlan1_qid, short mask) {
+    public static Criterion selectVlan1_Qid(long vlan1_qid, long mask) {
         return new Vlan1_QidCriterion(vlan1_qid, mask);
     }
 
-    public static Criterion selectVlan2_Tpid(short vlan2_tpid) {
+    public static Criterion selectVlan2_Tpid(long vlan2_tpid) {
         return new Vlan2_TpidCriterion(vlan2_tpid);
     }
   
-    public static Criterion selectVlan2_Tpid(short vlan2_tpid, short mask) {
+    public static Criterion selectVlan2_Tpid(long vlan2_tpid, long mask) {
         return new Vlan2_TpidCriterion(vlan2_tpid, mask);
     }
 
-    public static Criterion selectVlan2_Qid(short vlan2_qid) {
+    public static Criterion selectVlan2_Qid(long vlan2_qid) {
         return new Vlan2_QidCriterion(vlan2_qid);
     }
   
-    public static Criterion selectVlan2_Qid(short vlan2_qid, short mask) {
+    public static Criterion selectVlan2_Qid(long vlan2_qid, long mask) {
         return new Vlan2_QidCriterion(vlan2_qid, mask);
     }
 
-    public static Criterion selectDl_Type(short dl_type) {
+    public static Criterion selectDl_Type(long dl_type) {
         return new Dl_TypeCriterion(dl_type);
     }
   
-    public static Criterion selectDl_Type(short dl_type, short mask) {
+    public static Criterion selectDl_Type(long dl_type, long mask) {
         return new Dl_TypeCriterion(dl_type, mask);
     }
 
-    public static Criterion selectVer_Hl_E(Byte ver_hl_e) {
+    public static Criterion selectVer_Hl_E(long ver_hl_e) {
         return new Ver_Hl_ECriterion(ver_hl_e);
     }
   
-    public static Criterion selectVer_Hl_E(Byte ver_hl_e, Byte mask) {
+    public static Criterion selectVer_Hl_E(long ver_hl_e, long mask) {
         return new Ver_Hl_ECriterion(ver_hl_e, mask);
     }
 
-    public static Criterion selectTos_E(Byte tos_e) {
+    public static Criterion selectTos_E(long tos_e) {
         return new Tos_ECriterion(tos_e);
     }
   
-    public static Criterion selectTos_E(Byte tos_e, Byte mask) {
+    public static Criterion selectTos_E(long tos_e, long mask) {
         return new Tos_ECriterion(tos_e, mask);
     }
 
-    public static Criterion selectTot_Len_E(short tot_len_e) {
+    public static Criterion selectTot_Len_E(long tot_len_e) {
         return new Tot_Len_ECriterion(tot_len_e);
     }
   
-    public static Criterion selectTot_Len_E(short tot_len_e, short mask) {
+    public static Criterion selectTot_Len_E(long tot_len_e, long mask) {
         return new Tot_Len_ECriterion(tot_len_e, mask);
     }
 
-    public static Criterion selectIp_Id_E(short ip_id_e) {
+    public static Criterion selectIp_Id_E(long ip_id_e) {
         return new Ip_Id_ECriterion(ip_id_e);
     }
   
-    public static Criterion selectIp_Id_E(short ip_id_e, short mask) {
+    public static Criterion selectIp_Id_E(long ip_id_e, long mask) {
         return new Ip_Id_ECriterion(ip_id_e, mask);
     }
 
-    public static Criterion selectFrag_Off_E(short frag_off_e) {
+    public static Criterion selectFrag_Off_E(long frag_off_e) {
         return new Frag_Off_ECriterion(frag_off_e);
     }
   
-    public static Criterion selectFrag_Off_E(short frag_off_e, short mask) {
+    public static Criterion selectFrag_Off_E(long frag_off_e, long mask) {
         return new Frag_Off_ECriterion(frag_off_e, mask);
     }
 
-    public static Criterion selectTtl_E(Byte ttl_e) {
+    public static Criterion selectTtl_E(long ttl_e) {
         return new Ttl_ECriterion(ttl_e);
     }
   
-    public static Criterion selectTtl_E(Byte ttl_e, Byte mask) {
+    public static Criterion selectTtl_E(long ttl_e, long mask) {
         return new Ttl_ECriterion(ttl_e, mask);
     }
 
-    public static Criterion selectIpv4_E_Type(Byte ipv4_e_type) {
+    public static Criterion selectIpv4_E_Type(long ipv4_e_type) {
         return new Ipv4_E_TypeCriterion(ipv4_e_type);
     }
   
-    public static Criterion selectIpv4_E_Type(Byte ipv4_e_type, Byte mask) {
+    public static Criterion selectIpv4_E_Type(long ipv4_e_type, long mask) {
         return new Ipv4_E_TypeCriterion(ipv4_e_type, mask);
     }
 
-    public static Criterion selectIp_Check_E(short ip_check_e) {
+    public static Criterion selectIp_Check_E(long ip_check_e) {
         return new Ip_Check_ECriterion(ip_check_e);
     }
   
-    public static Criterion selectIp_Check_E(short ip_check_e, short mask) {
+    public static Criterion selectIp_Check_E(long ip_check_e, long mask) {
         return new Ip_Check_ECriterion(ip_check_e, mask);
     }
 
-    public static Criterion selectIp_Saddr_E(int ip_saIp_Saddr_Er_e) {
+    public static Criterion selectIp_Saddr_E(long ip_saIp_Saddr_Er_e) {
         return new Ip_Saddr_ECriterion(ip_saIp_Saddr_Er_e);
     }
   
-    public static Criterion selectIp_Saddr_E(int ip_saIp_Saddr_Er_e, int mask) {
+    public static Criterion selectIp_Saddr_E(long ip_saIp_Saddr_Er_e, long mask) {
         return new Ip_Saddr_ECriterion(ip_saIp_Saddr_Er_e, mask);
     }
 
-    public static Criterion selectIp_Daddr_E(int ip_daIp_Daddr_Er_e) {
+    public static Criterion selectIp_Daddr_E(long ip_daIp_Daddr_Er_e) {
         return new Ip_Daddr_ECriterion(ip_daIp_Daddr_Er_e);
     }
   
-    public static Criterion selectIp_Daddr_E(int ip_daIp_Daddr_Er_e, int mask) {
+    public static Criterion selectIp_Daddr_E(long ip_daIp_Daddr_Er_e, long mask) {
         return new Ip_Daddr_ECriterion(ip_daIp_Daddr_Er_e, mask);
     }
 
-    public static Criterion selectIpv6_Ver_Tp_Flb_E(int ipv6_ver_tp_flb_e) {
+    public static Criterion selectIpv6_Ver_Tp_Flb_E(long ipv6_ver_tp_flb_e) {
         return new Ipv6_Ver_Tp_Flb_ECriterion(ipv6_ver_tp_flb_e);
     }
   
-    public static Criterion selectIpv6_Ver_Tp_Flb_E(int ipv6_ver_tp_flb_e, int mask) {
+    public static Criterion selectIpv6_Ver_Tp_Flb_E(long ipv6_ver_tp_flb_e, long mask) {
         return new Ipv6_Ver_Tp_Flb_ECriterion(ipv6_ver_tp_flb_e, mask);
     }
 
-    public static Criterion selectIpv6_Plen_E(short ipv6_plen_e) {
+    public static Criterion selectIpv6_Plen_E(long ipv6_plen_e) {
         return new Ipv6_Plen_ECriterion(ipv6_plen_e);
     }
   
-    public static Criterion selectIpv6_Plen_E(short ipv6_plen_e, short mask) {
+    public static Criterion selectIpv6_Plen_E(long ipv6_plen_e, long mask) {
         return new Ipv6_Plen_ECriterion(ipv6_plen_e, mask);
     }
 
-    public static Criterion selectIpv6_E_Type(Byte ipv6_e_type) {
+    public static Criterion selectIpv6_E_Type(long ipv6_e_type) {
         return new Ipv6_E_TypeCriterion(ipv6_e_type);
     }
   
-    public static Criterion selectIpv6_E_Type(Byte ipv6_e_type, Byte mask) {
+    public static Criterion selectIpv6_E_Type(long ipv6_e_type, long mask) {
         return new Ipv6_E_TypeCriterion(ipv6_e_type, mask);
     }
 
-    public static Criterion selectIpv6_Hlmt_E(Byte ipv6_hlmt_e) {
+    public static Criterion selectIpv6_Hlmt_E(long ipv6_hlmt_e) {
         return new Ipv6_Hlmt_ECriterion(ipv6_hlmt_e);
     }
   
-    public static Criterion selectIpv6_Hlmt_E(Byte ipv6_hlmt_e, Byte mask) {
+    public static Criterion selectIpv6_Hlmt_E(long ipv6_hlmt_e, long mask) {
         return new Ipv6_Hlmt_ECriterion(ipv6_hlmt_e, mask);
     }
 
@@ -824,91 +824,91 @@ public final class Criteria {
         return new Ipv6_Dst_ECriterion(ipv6_dst_e, mask);
     }
 
-    public static Criterion selectUdp_Source(short udp_source) {
+    public static Criterion selectUdp_Source(long udp_source) {
         return new Udp_SourceCriterion(udp_source);
     }
   
-    public static Criterion selectUdp_Source(short udp_source, short mask) {
+    public static Criterion selectUdp_Source(long udp_source, long mask) {
         return new Udp_SourceCriterion(udp_source, mask);
     }
 
-    public static Criterion selectUdp_Dest(short udp_dest) {
+    public static Criterion selectUdp_Dest(long udp_dest) {
         return new Udp_DestCriterion(udp_dest);
     }
   
-    public static Criterion selectUdp_Dest(short udp_dest, short mask) {
+    public static Criterion selectUdp_Dest(long udp_dest, long mask) {
         return new Udp_DestCriterion(udp_dest, mask);
     }
 
-    public static Criterion selectLen(short len) {
+    public static Criterion selectLen(long len) {
         return new LenCriterion(len);
     }
   
-    public static Criterion selectLen(short len, short mask) {
+    public static Criterion selectLen(long len, long mask) {
         return new LenCriterion(len, mask);
     }
 
-    public static Criterion selectUdp_Check(short udp_check) {
+    public static Criterion selectUdp_Check(long udp_check) {
         return new Udp_CheckCriterion(udp_check);
     }
   
-    public static Criterion selectUdp_Check(short udp_check, short mask) {
+    public static Criterion selectUdp_Check(long udp_check, long mask) {
         return new Udp_CheckCriterion(udp_check, mask);
     }
 
-    public static Criterion selectSrv6_Type(Byte srv6_type) {
+    public static Criterion selectSrv6_Type(long srv6_type) {
         return new Srv6_TypeCriterion(srv6_type);
     }
   
-    public static Criterion selectSrv6_Type(Byte srv6_type, Byte mask) {
+    public static Criterion selectSrv6_Type(long srv6_type, long mask) {
         return new Srv6_TypeCriterion(srv6_type, mask);
     }
 
-    public static Criterion selectSrv6_Hdr_Ext_Len(Byte srv6_hdr_ext_len) {
+    public static Criterion selectSrv6_Hdr_Ext_Len(long srv6_hdr_ext_len) {
         return new Srv6_Hdr_Ext_LenCriterion(srv6_hdr_ext_len);
     }
   
-    public static Criterion selectSrv6_Hdr_Ext_Len(Byte srv6_hdr_ext_len, Byte mask) {
+    public static Criterion selectSrv6_Hdr_Ext_Len(long srv6_hdr_ext_len, long mask) {
         return new Srv6_Hdr_Ext_LenCriterion(srv6_hdr_ext_len, mask);
     }
 
-    public static Criterion selectSrv6_Routing_Type(Byte srv6_routing_Type) {
+    public static Criterion selectSrv6_Routing_Type(long srv6_routing_Type) {
         return new Srv6_Routing_TypeCriterion(srv6_routing_Type);
     }
   
-    public static Criterion selectSrv6_Routing_Type(Byte srv6_routing_Type, Byte mask) {
+    public static Criterion selectSrv6_Routing_Type(long srv6_routing_Type, long mask) {
         return new Srv6_Routing_TypeCriterion(srv6_routing_Type, mask);
     }
 
-    public static Criterion selectSrv6_Segments_Left(Byte srv6_segments_left) {
+    public static Criterion selectSrv6_Segments_Left(long srv6_segments_left) {
         return new Srv6_Segments_LeftCriterion(srv6_segments_left);
     }
   
-    public static Criterion selectSrv6_Segments_Left(Byte srv6_segments_left, Byte mask) {
+    public static Criterion selectSrv6_Segments_Left(long srv6_segments_left, long mask) {
         return new Srv6_Segments_LeftCriterion(srv6_segments_left, mask);
     }
 
-    public static Criterion selectSrv6_Last_Enty(Byte srv6_last_enty) {
+    public static Criterion selectSrv6_Last_Enty(long srv6_last_enty) {
         return new Srv6_Last_EntyCriterion(srv6_last_enty);
     }
   
-    public static Criterion selectSrv6_Last_Enty(Byte srv6_last_enty, Byte mask) {
+    public static Criterion selectSrv6_Last_Enty(long srv6_last_enty, long mask) {
         return new Srv6_Last_EntyCriterion(srv6_last_enty, mask);
     }
 
-    public static Criterion selectSrv6_Flags(Byte srv6_flags) {
+    public static Criterion selectSrv6_Flags(long srv6_flags) {
         return new Srv6_FlagsCriterion(srv6_flags);
     }
   
-    public static Criterion selectSrv6_Flags(Byte srv6_flags, Byte mask) {
+    public static Criterion selectSrv6_Flags(long srv6_flags, long mask) {
         return new Srv6_FlagsCriterion(srv6_flags, mask);
     }
 
-    public static Criterion selectSrv6_Tag(short srv6_tag) {
+    public static Criterion selectSrv6_Tag(long srv6_tag) {
         return new Srv6_TagCriterion(srv6_tag);
     }
   
-    public static Criterion selectSrv6_Tag(short srv6_tag, short mask) {
+    public static Criterion selectSrv6_Tag(long srv6_tag, long mask) {
         return new Srv6_TagCriterion(srv6_tag, mask);
     }
 
@@ -936,35 +936,35 @@ public final class Criteria {
         return new Srv6_Segmentlist3Criterion(srv6_segmentlist3, mask);
     }
 
-    public static Criterion selectIpv6_Ver_Tp_Flb_I(int ipv6_ver_tp_flb_i) {
+    public static Criterion selectIpv6_Ver_Tp_Flb_I(long ipv6_ver_tp_flb_i) {
         return new Ipv6_Ver_Tp_Flb_ICriterion(ipv6_ver_tp_flb_i);
     }
   
-    public static Criterion selectIpv6_Ver_Tp_Flb_I(int ipv6_ver_tp_flb_i, int mask) {
+    public static Criterion selectIpv6_Ver_Tp_Flb_I(long ipv6_ver_tp_flb_i, long mask) {
         return new Ipv6_Ver_Tp_Flb_ICriterion(ipv6_ver_tp_flb_i, mask);
     }
 
-    public static Criterion selectIpv6_Plen_I(short ipv6_plen_i) {
+    public static Criterion selectIpv6_Plen_I(long ipv6_plen_i) {
         return new Ipv6_Plen_ICriterion(ipv6_plen_i);
     }
   
-    public static Criterion selectIpv6_Plen_I(short ipv6_plen_i, short mask) {
+    public static Criterion selectIpv6_Plen_I(long ipv6_plen_i, long mask) {
         return new Ipv6_Plen_ICriterion(ipv6_plen_i, mask);
     }
 
-    public static Criterion selectIpv6_I_Type(Byte ipv6_i_type) {
+    public static Criterion selectIpv6_I_Type(long ipv6_i_type) {
         return new Ipv6_I_TypeCriterion(ipv6_i_type);
     }
   
-    public static Criterion selectIpv6_I_Type(Byte ipv6_i_type, Byte mask) {
+    public static Criterion selectIpv6_I_Type(long ipv6_i_type, long mask) {
         return new Ipv6_I_TypeCriterion(ipv6_i_type, mask);
     }
 
-    public static Criterion selectIpv6_Hlmt_I(Byte ipv6_hlmt_i) {
+    public static Criterion selectIpv6_Hlmt_I(long ipv6_hlmt_i) {
         return new Ipv6_Hlmt_ICriterion(ipv6_hlmt_i);
     }
   
-    public static Criterion selectIpv6_Hlmt_I(Byte ipv6_hlmt_i, Byte mask) {
+    public static Criterion selectIpv6_Hlmt_I(long ipv6_hlmt_i, long mask) {
         return new Ipv6_Hlmt_ICriterion(ipv6_hlmt_i, mask);
     }
